Fixed more bugs with multiplayer
Fixed a bug that gives an exception error when the directory in a character is not set correctly. Now it will send a warning in the logs and ignore the directory when giving results.
Fixed doneNote not turning on when starting
Fixed Dreadbear's directory saying "Dreabear" instead of "Dreadbear"
Fixed a bug where gameManger got set to null in game.cs once a game has finished.
Added a check to Leave if the player object still exists in case the host disconnected.
Fixed a bug where if you're not the host you can't see the result.